AuthorInfo extracted from AuthorList
